% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.quantileplot.R
\name{plot.quantileplot}
\alias{plot.quantileplot}
\title{Plot function for \code{quantileplot} objects}
\usage{
\method{plot}{quantileplot}(x, bottom_xlab_angle = NULL, bottom_height = 0.3, ...)
}
\arguments{
\item{x}{An object of class \code{quantileplot}, which results from a call to the function \code{quantileplot}}

\item{bottom_xlab_angle}{Numeric degrees to rotate x-axis labels}

\item{bottom_height}{Proportion of vertical space in the full plot to be taken up by uncertainty draws at the bottom (only relevant if the \code{quantileplot} includes uncertainty draws).}

\item{...}{Other arguments to \code{plot} commands}
}
\value{
Prints a plot. If the \code{quantileplot} was specified with option \code{uncertainty_draws = TRUE}, this is a \code{gtable} object with a panel of plots. If the \code{quantileplot} was specified with \code{uncertainty_draws = FALSE} (the default), this is a \code{ggplot2} object.
}
\description{
Plots a \code{quantileplot} object.
}
\references{
Lundberg, Ian, Robin C. Lee, and Brandon M. Stewart. 2021. "The quantile plot: A visualization for bivariate population relationships." Working paper.

Lundberg, Ian, and Brandon M. Stewart. 2020. "Comment: Summarizing income mobility with multiple smooth quantiles instead of parameterized means." Sociological Methodology 50(1):96-111.

Fasiolo, Matteo, Simon N. Wood, Margaux Zaffran, Raphaël Nedellec, and Yannig Goude. 2020. "Fast calibrated additive quantile regression." Journal of the American Statistical Association.
}
