% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.quantileplot.R
\name{print.quantileplot}
\alias{print.quantileplot}
\title{Print function for \code{quantileplot} objects}
\usage{
\method{print}{quantileplot}(x, ...)
}
\arguments{
\item{x}{An object of class \code{quantileplot}, which results from a call to the function \code{quantileplot}}

\item{...}{Other arguments to \code{print} commands}
}
\value{
Prints the plot generated by \code{quantileplot}
}
\description{
Prints the plot generated by \code{quantileplot}
}
\references{
Lundberg, Ian, Robin C. Lee, and Brandon M. Stewart. 2021. "The quantile plot: A visualization for bivariate population relationships." Working paper.

Lundberg, Ian, and Brandon M. Stewart. 2020. "Comment: Summarizing income mobility with multiple smooth quantiles instead of parameterized means." Sociological Methodology 50(1):96-111.

Fasiolo, Matteo, Simon N. Wood, Margaux Zaffran, Raphaël Nedellec, and Yannig Goude. 2020. "Fast calibrated additive quantile regression." Journal of the American Statistical Association.
}
