% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.quantileplot.R
\name{summary.quantileplot}
\alias{summary.quantileplot}
\title{Summary function for \code{quantileplot} objects}
\usage{
\method{summary}{quantileplot}(object, ...)
}
\arguments{
\item{object}{An object of class \code{quantileplot}, which results from a call to the function \code{quantileplot}}

\item{...}{Other arguments to \code{summary} commands}
}
\value{
Prints a summary of the estimates.
}
\description{
Summarizes the S3 class object returned by the quantileplot function
}
\references{
Lee, Robin C., Ian Lundberg, and Brandon M. Stewart. 2021. "Smooth quantile visualizations enhance understanding of bivariate population distributions." Working paper.

Lundberg, Ian, and Brandon M. Stewart. 2020. "Comment: Summarizing income mobility with multiple smooth quantiles instead of parameterized means." Sociological Methodology 50(1):96-111.

Fasiolo, Matteo, Simon N. Wood, Margaux Zaffran, Raphaël Nedellec, and Yannig Goude. 2020. "Fast calibrated additive quantile regression." Journal of the American Statistical Association.
}
