% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_curves.R
\name{gen_curves}
\alias{gen_curves}
\title{Generate Smooth Curves for a \code{quantileplot}}
\usage{
gen_curves(
  formula,
  second_formula,
  data,
  weights = NULL,
  quantiles = c(0.1, 0.25, 0.5, 0.75, 0.9),
  show_ci = FALSE,
  ci = 0.95,
  uncertainty_draws = 10,
  ...
)
}
\arguments{
\item{formula}{A bivariate model formula (e.g. \code{y ~ x})}

\item{second_formula}{Model formula to allow the learning rate to change as a function of the predictor. This is passed to \code{mqgam} as the second element in the \code{form} argument. Defaults to the same specification as \code{formula} but without the outcome variable.}

\item{data}{Data frame containing the variables in \code{formula}. If \code{weights} are specified, they must be a column of \code{data}.}

\item{weights}{String name for sampling weights, which are a column of \code{data}. If not given, a simple random sample is assumed.}

\item{quantiles}{Numeric vector containing quantiles to be estimated. Values should be between 0 and 1.}

\item{show_ci}{Logical, defaults to \code{FALSE}. Whether to show credible intervals for the estimated smooth quantile curves.}

\item{ci}{Numeric probability value for credible intervals; default to 0.95 to produce 95 percent credible intervals. Only relevant if \code{show_ci = TRUE}.}

\item{uncertainty_draws}{Numeric. If non-null, the number of simulated posterior draws to estimate for each smooth quantile curve. When used with the \code{plot} function, these appear in panels below the main plot.}

\item{...}{Other arguments passed to \code{mqgam} for fitting of smooth quantile curves.}
}
\value{
A list of length 2. Element \code{curves} is a data frame containing the data for plotting smooth curves for quantiles of the outcome given the predictor. Element \code{mqgam.out} is the fitted object from \code{mqgam}.
}
\description{
Estimate smooth curves for quantiles of the outcome as a function of the predictor. This function calls \code{mqgam} from the \code{qgam} package. This function is typically called indirectly via a user call to \code{\link{quantileplot}}.
}
\references{
Lee, Robin C., Ian Lundberg, and Brandon M. Stewart. 2021. "Smooth quantile visualizations enhance understanding of bivariate population distributions." Working paper.

Lundberg, Ian, and Brandon M. Stewart. 2020. "Comment: Summarizing income mobility with multiple smooth quantiles instead of parameterized means." Sociological Methodology 50(1):96-111.

Fasiolo, Matteo, Simon N. Wood, Margaux Zaffran, Raphaël Nedellec, and Yannig Goude. 2020. "Fast calibrated additive quantile regression." Journal of the American Statistical Association.
}
